requirements.txt
# Core
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
google-genai>=1.0.0
pydantic>=2.0.0

# Data Processing
pandas>=2.0.0

# External APIs
finnhub-python>=2.4.0
aiohttp>=3.9.0

# Utilities
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
